// src/air/interfaces/sidoName.type.ts
export interface SidoName {
  sidoName:
    | '전국'
    | '서울'
    | '부산'
    | '대구'
    | '인천'
    | '광주'
    | '대전'
    | '울산'
    | '경기'
    | '강원'
    | '충북'
    | '충남'
    | '전북'
    | '전남'
    | '경북'
    | '경남'
    | '제주'
    | '세종';
}